Make required changes.
