+++ person.rb
@@ -1,0 +1,22 @@
+class Person
+  attr_accessor :name, :age
+
+  attr_reader :id
+
+  def initialize(age, name = 'unknown', parent_permission: true)
+    @name = name
+    @age = age
+    @parent_permission = parent_permission
+    @id = 0
+  end
+
+  def can_use_services
+    of_age || @parent_permission
+  end
+
+  private
+
+  def of_age
+    @age >= 18
+  end
+end
